pyproject.toml
[project]
name = "sbv2-dataset-editor-private"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altgraph==0.17.5",
    "asttokens==3.0.1",
    "auto-py-to-exe==2.48.1",
    "bottle==0.13.4",
    "bottle-websocket==0.2.9",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "eel==0.18.2",
    "executing==2.2.1",
    "future==1.0.0",
    "gevent==25.9.1",
    "gevent-websocket==0.10.1",
    "greenlet==3.3.0",
    "icecream==2.1.8",
    "idna==3.11",
    "importlib-resources==6.5.2",
    "numpy==2.4.0",
    "packaging==25.0",
    "pefile==2024.8.26",
    "pillow==12.0.0",
    "pycparser==2.23",
    "pygments==2.19.2",
    "pyinstaller==6.17.0",
    "pyinstaller-hooks-contrib==2025.11",
    "pyparsing==3.3.1",
    "pyperclip==1.11.0",
    "pywin32-ctypes==0.2.3",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "setuptools==80.9.0",
    "sounddevice==0.5.3",
    "soundfile==0.13.1",
    "tkeasygui==1.0.40",
    "typing-extensions==4.15.0",
    "urllib3==2.6.2",
    "zope-event==6.1",
    "zope-interface==8.1.1",
]
